requirements.txt
boto3==1.9.53
botocore==1.12.53
certifi==2018.10.15
chardet==3.0.4
configparser==3.5.0
contextlib2==0.5.5
Cython==0.25.2
dj-database-url==0.5.0
Django==2.1.3
django-heroku==0.3.1
docutils==0.14
future==0.17.1
gunicorn==19.9.0
h5py==2.8.0
idna==2.7
imageio==2.4.1
jmespath==0.9.3
line-bot-sdk==1.8.0
nnabla==1.0.9
numpy==1.15.4
onnx==1.3.0
Pillow==5.3.0
protobuf==3.6.1
psycopg2==2.7.6.1
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
requests==2.20.1
s3transfer==0.1.13
scipy==1.1.0
six==1.11.0
tqdm==4.28.1
typing==3.6.6
typing-extensions==3.6.6
urllib3==1.24.1
whitenoise==4.1.2
wincertstore==0.2
